More work. Fix core tests.
