+++ packages/core/src/Adapter.tsx
@@ -1,6 +1,6 @@
 import uuid from 'uuid/v4';
 import { isObject, stripClassPrefix } from 'aesthetic-utils';
-import aesthetic from './instance';
+import Aesthetic from './Aesthetic';
 import CacheManager from './CacheManager';
 import Sheet from './Sheet';
 import StyleSheetManager from './StyleSheetManager';
@@ -12,11 +12,13 @@ export default abstract class Adapter<
   NativeBlock extends object,
   ParsedBlock extends object | string = NativeBlock
 > {
+  aesthetic!: Aesthetic;
+
+  syntax = new UnifiedSyntax<NativeBlock>();
+
   protected cacheManager = new CacheManager<SheetMap<ParsedBlock>>();
 
   protected sheetManager: StyleSheetManager | null = null;
-
-  protected syntax = new UnifiedSyntax<NativeBlock>();
 
   /**
    * Apply and inject global styles for the current theme.
@@ -26,7 +28,7 @@ export default abstract class Adapter<
     // Templates may be generated by Webpack or something similar,
     // so set this programmatically based on the defined option.
     if (typeof document !== 'undefined') {
-      document.documentElement.setAttribute('dir', aesthetic.options.rtl ? 'rtl' : 'ltr');
+      document.documentElement.setAttribute('dir', this.aesthetic.options.rtl ? 'rtl' : 'ltr');
     }
 
     const options = this.prepareTransformOptions({
@@ -39,7 +41,7 @@ export default abstract class Adapter<
     delete options.dir;
 
     const cache = this.cacheManager.get(GLOBAL_STYLE_NAME, options);
-    const globalSheet = aesthetic.getGlobalSheet(options.theme);
+    const globalSheet = this.aesthetic.getGlobalSheet(options.theme);
 
     if (cache || !globalSheet) {
       return this;
@@ -76,7 +78,7 @@ export default abstract class Adapter<
     this.applyGlobalStyles(baseOptions);
 
     const nativeSheet = this.syntax.convertStyleSheet(
-      aesthetic.getStyleSheet(styleName, options.theme),
+      this.aesthetic.getStyleSheet(styleName, options.theme),
       {
         ...options,
         name: styleName,
@@ -100,6 +102,26 @@ export default abstract class Adapter<
    * If no target defined, will flush all buffered styles.
    */
   flushStyles(styleName?: StyleName) {}
+
+  /**
+   * Return the standard cache manager.
+   */
+  getCacheManager(): CacheManager<SheetMap<ParsedBlock>> {
+    return this.cacheManager;
+  }
+
+  /**
+   * Return a native style sheet manager used for injecting CSS.
+   */
+  getStyleSheetManager(): StyleSheetManager {
+    if (this.sheetManager) {
+      return this.sheetManager;
+    }
+
+    this.sheetManager = new StyleSheetManager();
+
+    return this.sheetManager;
+  }
 
   /**
    * Return true if the style object is a parsed block and not a native block.
@@ -200,29 +222,16 @@ export default abstract class Adapter<
   abstract transformToClassName(styles: ParsedBlock[]): ClassName;
 
   /**
-   * Return a native style sheet manager used for injecting CSS.
-   */
-  protected getStyleSheetManager(): StyleSheetManager {
-    if (this.sheetManager) {
-      return this.sheetManager;
-    }
-
-    this.sheetManager = new StyleSheetManager();
-
-    return this.sheetManager;
-  }
-
-  /**
    * Return transform options with defaults applied.
    */
   protected prepareTransformOptions(options: TransformOptions = {}): Required<TransformOptions> {
-    const dir = aesthetic.options.rtl ? 'rtl' : 'ltr';
+    const dir = this.aesthetic.options.rtl ? 'rtl' : 'ltr';
 
     return {
       dir: options.dir || dir,
       global: options.global || false,
       name: options.name || '',
-      theme: options.theme || aesthetic.options.theme,
+      theme: options.theme || this.aesthetic.options.theme,
     };
   }
 }
