+++ packages/core/src/index.ts
@@ -3,8 +3,9 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-import aesthetic from './instance';
+import instance from './instance';
 import Adapter from './Adapter';
+import Aesthetic from './Aesthetic';
 import ClassNameAdapter from './ClassNameAdapter';
 import UnifiedSyntax from './UnifiedSyntax';
 import Ruleset from './Ruleset';
@@ -13,6 +14,6 @@ import Sheet from './Sheet';
 export * from './constants';
 export * from './types';
 
-export { Adapter, ClassNameAdapter, UnifiedSyntax, Ruleset, Sheet };
+export { Adapter, Aesthetic, ClassNameAdapter, UnifiedSyntax, Ruleset, Sheet };
 
-export default aesthetic;
+export default instance;
