--- packages/core/tests/ClassNameAesthetic.test.ts
@@ -1,13 +1,0 @@
-import ClassNameAesthetic from '../src/ClassNameAesthetic';
-
-describe('ClassNameAesthetic', () => {
-  let instance: ClassNameAesthetic<{}>;
-
-  beforeEach(() => {
-    instance = new ClassNameAesthetic();
-  });
-
-  it('joins and returns class names', () => {
-    expect(instance.transformStyles(['foo', 'bar'], {})).toBe('foo bar');
-  });
-});
